202.2. LoadMore component - fetching next comments (different way of pagination)(Section 21: React Redux Application)
So let's modify our comment to fetch action creator to also specify a page to load and set a default

to 1.

Now our URL is blog posts slash ID which is the blog post these comments and we have to pass the page

parameters so we have underscore page because that's in our API platform project.

But if you modify it be sure to use the name that you changed it to.

Now you can see that by default we are starting with fetching the comments from the page one.

So even if you want specify the page the request is being made to that you are well we have the page

parameter.

Now our producer needs a default state for our nation.

The current page will be one and the page count will be known Same like with the blog post.

These and mine that they are just substate So our global state contains comment list item and block

post list item and they have their own copies of this.

All of this parameter so the current page and page counts so you are not modifying their other state

or the other page of the blog posts so this is a separate state.

Now the comment is received.

Reduce should modify the current page statically so it should just get the current page from state and

the one to so well the current page would be the next page.

So each time this button will be pressed we will load it then next page and the page count.

We can use our useful function from the API.

It is the Heidrek page count because the return data format is the same and also very important to do.

When something bad happens or the comment list page isn't loaded so the component is unmounted we have

to reset the page to one and the page count to learn not to show comments from a different page on a

completely different block post.

So you have to modify the state explicitly nothing happens out magic outer magically.

Now let's go back to our comment list.

Container.

So we have to figure out when we should show this load more button because it shouldn't be shown.

Always It's only when there is more than one page of the comments so we will have to verify that.

Currently we are rendering.

Always.

But now let's add a simple check so we will do a conditional rendering so only if the first name only

first condition is met.

Well basically whenever you change all these conditions and the final after this ampersand sign you

will render the component.

It will only render when all of these conditions are met.

So we are checking the page count is more than one and the current page is less or equal than the page

count.

So this makes sense for this component because it has 15 comments but if you will find another blog

post with two comments or anyway less than 10.

This pattern won't be shown so this condition works great now.

Well we can move that condition to OC to a variable.

So this would be clear for anyone.

What are we doing here.

And let's just move it to a variable called shock load more so this will be Combest so it won't change

any time so it's just a constant basically can't be changed won't change but maybe it will be clear

for people looking at this code.

What it is what is happening and what are those conditions.

So we have our conditional rendering.

And now since this is a dummy component we need the click method.

So we will tell this load more component.

What should happen after it is being clicked.

So let's define own load more click metadata inside our comment list container

and we will get or structure some of our props so we will.

We are getting the blog post ID and the current page and what we are going to do is we will dispatch

our comment leads to fetch which is responsible for fetching the comments from API and will pass the

blog post ID which it requires So it does require parameter.

And the second one will be the current page so mind that each time this is being dispatched the comment

list fetch the current page changes inside the state automatically.

And so when we press this button we can safely just use the current page from the state because it is

already modified after the previous fetching of the comments has finished.

So even if this page is loaded for the first time it will just move the current page to 2 so the next

click would load the comments from the page to.

Now we are putting that disabled and on the click as props to this lot more so we can disable this button

while the comments are fetching.

So this is passed through props and also we are passing the on click handler.

So also this component would not be aware of what it's doing.

It will just have the event handler on Click now.

Let's see if that works so well kind of loaded next comments but unfortunately it just removed the comments

we had before we should just append them to their initial comments and just replace them.

So we still need to work on it and we will do that in the next lecture.