203. Append new comments inside reducer(Section 21: React Redux Application)
So last time we other the load more bottom to load more comments when there is more than 10 comments

and we had a problem that the data gets gets replaced by the next page data.

So now the fix is very simple we just have to modify the reducer So when handling our comment least

received action instead of just blindly assigning Whatever comes back from the response which is inside

the Heidrun member we should first check if the state comments is already neutralized because it would

be if we fed it for the first time.

So by using the negation we check if it's no or undefined if it is then we just assign whatever came

back from that action.

But if it is already initialized so there is some initial comments that would mean that the first page

at least was first then seems comments state comments on an array of elements we can use concat on that.

And this would return in the array itself with the added elements and we'll add the elements returned

from the API.

So apparently it didn't work correctly.

And well yes the problem was that we used the wrong wrong state properties so it wasn't comments it

was comments list.

So let's start now.

Load more.

It still doesn't work exactly as we would like it to.

It did load and the remaining comments.

But the problem was that it just showed the spinner.

So that's not what we might have wanted.

It just all the comments disappeared and then appeared again with something that was added.

So it's very unclear for the user what happened.

So I would get lost and e-filed been reading some comments and then like a lot more.

And suddenly everything would disappear and reload.

I would totally lost lose it when I was reading it and what actually happened which are new which are

old and what page I'm on.

So you can fix that and only render the spinner if the current page is the first page.

So the spinner would only be shown forward when the page initially load.

So now it's showing because we are on the first page but now load comments well you can see it perfectly

animated.

Adding the remaining comments and the button has disappeared because there is no more pages.