205.1. Registration action creators and form validation(Section 21: React Redux Application)
So in this lecture we will actually implement all the logic related to registration form.

So this lecture will be a little bit longer but well I didn't want to split that.

So let's start with setting that button as disabled while the form is submitting.

So the Submitting is our property we get from redux form to our props so currently obviously doesn't

work because our form has to have the handles submit first because otherwise it is just behaving as

a typical form.

So wrapping this component with redox form does not make our form inside out actually being handled

by the Rydex forum.

So as usual we have to handle subnet that will accept itself and mettled which is ours in which we will

implement the metal.

The action dispatching to the back.

Now let's see if we can send the data.

So let's check out the console we are logging it and it looks that we send the values properly.

Now we have our check box that you have to accept before you can even send the form.

And as I said last time we'll be handling that using this state.

So state should be the components state should be realized in the constructor.

So we are giving it an initial stage.

So mine is different from the Rydex stayed the redox state is the global state of the application that

has nothing to do with components.

It is just bound to components using our Hendee metals and we are just passing the state to the props

using redux.

And the state that we've added in the constructor is basically the state of this individual component

which will also handle the real rendering of the component when it changes and it can be modified.

Well it can be neutralized using just simple assignment.

So let's see how that works.

So we have our own terms accepts that Cliq click handler that we've assigned to the On Click on that

or on our checkbox.

Now to modify this state we cannot just assign this state again like we assigned to our property.

We have to call this said state otherwise react we'll just ignore that wouldn't.

Recall that the state has changed.

So this state accepts a metal which gets passed the previous state so we can for example compare it.

And we are doing this right now so we are modifying the terms accept that property of the state by negating

the previous terms accept that state.

So it's pretty simple and you can see that in the console that each time we clicked the current state

has the terms accepted with either false or true.

So this is a pretty simple thing.

We haven't shown this state previously.

We started opposite of what most of us are doing starting with this state.

We started straight with the Rydex But this is how you will how you can keep state on the components.

Now we've modified the disabled of the bottom to only let users submit the form if the form is not submitting

already or the terms are accepted.

So those two values has to be true or well at least the form has to form submitting has to be false

and the terms accepted has to be true.

Otherwise the button is disabled.

And this is what we expected.