205.2. Registration action creators and form validation(Section 21: React Redux Application)
Now let's handle our actual creators because we need to somehow send this data to the back and we can

copy the user logging attempt action creator because well it would be pretty similar.

So let's not rewrite it again you can of course do that for practice so I'd suggest you just stop the

video right now and try to type it.

It's always a good exercise and we will be sending our registration data to the users.

And so this would be post to the users and we need the username password retaped password name to be

sent.

So we are passing all those as a request body minds that we are just using the variable names because

there is this trick that when you just used a variable name it adds a property to the object with this

name and the value of that property now different to the user logging.

We won't be dispatching any action on success currently will only handle that errors.

So remember that we have this bar's API error order.

And now it is a perfect case for using this function because the form is pretty big.

We have lots of different fields on it.

So mind that for catch we'll have to pass that error because it contains the response so we have to

really have to have those errors back from the response to parse them.

So basically we will be only showing their errors on the form and the form would be successful.

We will just re direct.

Now we'll need that connect from re-act redux to easily dispatch actions from our own submit so we can

easily send this form to the endpoint.

So we have important action creator.

And now let's drop that component using conect but we only have mapped dispatched to prop's So we are

not mapping state the prop's now how we can send this form easily.

So you can either supply all the properties all the values which are the fields on the values because

values is are plain javascript object all or alternatively you can convert their values to an array

using object values and then pass it to the method argument using free dots which will just like and

packet to the list of arguments.

So if you cancel look the result of this you can see that we have five individual values.

So we'll just use it as a shorthand for our big object.

Of course you are free to just pass every parameter individually by using the values dot name of the

parameter in the correct order.

Now when we submit this form and it ends up with success mind that we weren't dispatching any additional

actions then we'll just reset the form and read direct to the main page.

So let's see our agree button works.

And the sending.

Also works perfectly.

We have all the errors.

Well not perfectly but we'll see in a bit what doesn't work.

So let's try registering as an admin.

So in this case it didn't work because it has to be these six characters so we have this user name only

because we added that fixtures but maybe let's try John Doe and there is no error under a user name

and that should be because this user already exists.

So it seems we have something wrong inside our IP platform and that is our unique entity because we

have specified that the validation group for sending a new user is post.

You can even see that on the top of the screen validation groups post.

So when you have a validation group set for a certain operation you have to specify those group groups

on all the constraints that you would like to use.

Otherwise it won't be used by default.

It is always a default group so if there is anything as specified as a validation group it just won't

work.

Now you can see these values already used under the user name so it is validating.

So let's find a user name that does not exist.

That would be Jondo to now type our password that will be correct.

You can always validate that during writing.

So each time you press register you will get errors of what's not correct yet.

And we can use my username so we can see that button is disabled.

And after success we get immediately redirected and we get back to the user.

And when you refresh the ask you a light you can see that there is John Doe to create that which is

a commentator and has a confirmation token.

So we should now be able to confirm this user and we'll be also implementing that in react so we'll

be creating forums for typing.

Consumation talkin received through their e-mail.