207. RegistrationContainer - 2 step registration process(Section 21: React Redux Application)
So last time we created a registration form and we have received that token now we have to handle it

using the confirmation token forms so let's create this new component.

Call it confirmation form.

Yes.

And this would also be another form to which the user can send the token and confirm his account immediately

after registering.

So this would be just another way this can be done because we also have this traditional Symphonie controller

with the link inside the email.

So we create our confirmation form the usual way.

So we are importing to react and to the field and relax form.

So create those confirmation form extend the reactor component and now implementing this form will be

pretty simple because it's not much different from what we have inside our registered forms so we have

to wrap that component with the Rydex form.

Give this forum a name so let's let it be confirmation form and now we can even copy the code that we

have for the registered form because we'll be using the same styling.

So maybe not everything is useful.

We can for example copy the whole return that returns the JSE X and then only modify it to what we will

actually need.

So we have this big form.

We have the handle submit and submitting to see that the form is currently submitting and we'll remove

all the fields except the one field on the top and change the button labeled to confirm your account

and also remove this check for the terms accepted and modify the field name and label to confirmation

token.

It might also be useful to give some clear message that this is to be used for confirming the talkin.

So let's right.

Please confirm your account with and you received in your e-mail

OK.

And now finally let's add the confirmation form to the redux forum Cole

now we will be creating a new container component that will contain the registration form itself and

the confirmation form

so it will work the way that we have the registration form and when the user successfully registers

We will show him their confirmation form.

So we are creating a new registration container component and it should simply return register form.

So for the for the start for the beginning it will just render the register form and that's it will

be later implementing all the logic that will render the specific form depending on what the current

state in the user registration.

So inside the app let's switch the register form with the registration container so we can remove the

import statement.

And this is how our register link looks like so we didn't change but we will be implementing that logic

that would make make it showed that confirmation for them after you initially sent the registration

form.