208. Showing ConfirmationToken form on successful account creation(Section 21: React Redux Application)
So last time we created registration container component that only currently contains the register form.

Now let's create new reducers So we will be able to keep track of the state of the registration process.

So this state will contain two properties.

And this would be a registration success and confirmation of success.

And they both default to false because obviously before we accomplish anything they are both by default

false.

Now let's put this reducer to all of our our producers using combine reducers.

Now we have to wrap this in this container component using mops.

They do props because we would like to be able to read their state of the registration process easily.

So we are using their connect from Rydex.

So currently we only rendered the register form and if we will have this state we can now conditionally

render the appropriate step.

So we are getting the registration success and confirmation success from this process which is being

up from their global Up-State or the actual registration substate.

Now we should render that register form only if the registration success is yet falls.

So this would mean that no step was finished

and we had an error because we needed to import the component again because previously we were exporting

it to using export default.

And now we modified the way we export it.

So it's no longer default or it is default.

So actually it is now default.

So we had to modify the imports statement.

Now we need to handle somehow that the user properly registered.

So we need to add a new action creator and maybe call it the user register a success and this will be

dispatched when the user registers or when the request to post user is successful.

Obviously we need a constant for that.

So let's get to Konstanz.

Yes and now import it imported inside our action create our.

So it is on the bottom of our import least

So when we do this budget this auction that registration was successful now we can modify the state

inside our registration reduce our.

So when this happens the user registers success their registration success should change to true.

So this is our first step.

And if the registration success would be true we'll just go to another step so this if statement would

be false and then we can go to the next step and the next step would be to check if the confirmation

success is false because if it does that means we have to return our confirmation form.

So we should be able to validate this right now.

So we are removing a user I am removing a user we have to use it with the email I use to receive the

confirmation tokens so I can register again.

And also inside the register form I am removing the re direct so it won't really direct the user to

the main page of our up after registering because now this is being handled inside a registration container

and the registration container is responsible for displaying first the register form and after registering

the confirmation talkin form.

So we have that on submit confirmation form meathook but we won't implement it right now.

Now let's create a user with a valid email so we can get the talken.

So you see that the registration was successful and now we have our confirmation form.

And in the next lecture we actually implement it so it can send in accept their confirmation talk and

to verify the user.