209. Keeping state of registration/confirmation process(Section 21: React Redux Application)
So last time we were able to create a user account and seen the confirmation took in form but we haven't

implemented the sending of those token to our buchen so it can be verified.

So we will do this inside this lecture so maybe we can start by copying the user register action creator

and modify its arguments because this actually accepts only a confirmation token and we will send this

confirmation token but not to the users and point.

But we can check in our docs it is users slash confirm.

So users slash confirm is there you are for our custom user configuration resource.

It accepts only post

and now we will also create another action creator to signal that the confirmation was successful.

So we are doing the same thing as we did with the registration itself but this time we were dispatched

that the confirmation process was successful.

So later in our producer we'll be able to set the flag for confirmation success to true.

So we need to.

As usual the constant for this action and on imports so I'm doing this using that BHP storm import feature.

But if you have a different editor you will have to add this import manually.

Now let's take a look at our user Confiteor method.

So it looks pretty correct.

Maybe except the error handling but we'll check that later.

So inside our confirmation form now we'll have to handle the sending of that form.

So we need to use Rydex to map our dispatching dispatching of our actions to the props of this component

so we can easily call this method to send the request to confirm a user in our own submit.

So a dispatched to prop's object import and action creator function and also wrap the confirmation form

with the conect

and our two arguments.

First is the map state of props and second map dispatched the props since we are not mapping and state

inside this component.

We only have moppin dispatch to prop's

now will also copy the same handler that we had inside the confirmation sorry the registration form.

But we will simply call the user confirm instead of user register and since we only have one field and

there is one argument to use or confirm we can just get the value from the values object.

And we also reset the form when it is being submitted.

So orders sorry when it is actually successful so when the form is sent successfully and the rest response

is two hundred or two hundred one.

And the final step is to modify this date inside the reducer.

So when we see the confirmation success action we modify that confirmation Sachs's to true.

And this would be the last step in our registration process.

And after both of those steps are successful we'll be able to display the user a kind of message that

his account is created and confirmed.

So we do not have the files that render.

We'll work on this later.

So we have our End Point.

It looks like this inside their documentation.

So now let's go to our API platform project so we can see what happens when the token is incorrect so

it does not have validation exceptions so it won't return.

Nice response with all the constraint violations so we shouldn't use parse API errors in this case.

We should just return our generic error.

So this is by we can do that by returning object with underscore error.

And this generic error message could then be rendered on the form.