210.1. Registration/Confirmation process, with redirect timer(Section 21: React Redux Application)
So let's continue working on our registration process.

And the final step in this registration container is to display our success message when the user has

successfully registered and then provided a valid confirmation.

So now we are certain that his account is properly created and also confirmed.

So only then the user would be able to log in so initially we'll just render a simple box as we did

all the time when we have for example block post with the congratulations message.

But we will also like to create a counter which will say that user will be redirected in a couple of

seconds to the main page so he can log in but we'll do that in the coming couple of minutes.

So first we will also create another ruction action creator that will signal that the registration process

is completed.

Or actually it might also signal that the process was interrupted.

So we only allow users to immediately confirm their token after they have registered.

Otherwise they'll have to use the link from the e-mail.

So we will do that that when the user started the registration process created an account but haven't

confirmed an account and then went to another page.

We need a way to reset all this data that we have in the registration reduce.

So the registration complete well maybe it should also be called user register process reset.

And because this will be used both when that process is successful and is completed fully by the user.

And when that process is somehow interrupted so we will reset this in the registration container.