210.2. Registration/Confirmation process, with redirect timer(Section 21: React Redux Application)
As I said and

we can handle that using that component will unmount.

So this is our lifecycle matter that will be cold when the component will be just removed wouldn't be

rendered anymore.

And this is always the case when you change the route so different component is being rendered.

So that's why we need to map this patch to prop's So any time someone will not finish the process started

or just go for one step that those flock's for the registration process will be reset.

Now to have this counter that will count from 10 seconds or any other number that you can think of might

be only five seconds we will actually keep.

Also the state on the registration container component so we stayed inside the constructor.

It's always a good place to utilize the state of the component inside constructor.

And what we will do is we will have a component Dete update.

And in this method we will check if the props have changed.

So we'll check if the confirmation success is different than it was previously.

And also whether it is true.

So it can only be true when the user has already registered so registrations access is true and then

has confirmed his account.

So only then this condition will be true.

And when this happens we'll set a timer so we'll create a timer that will call the same function.

Each second and seemed to initialize a timer we have to clear it.

So we have this interval that we'll call this function that we haven't implemented yet every second.

And when that component will mount we need to clear this interval.

So it will be stored inside the timer field of this component.

And the method that will be called in this one second interval is very simple.

It will just modify the state of this component so don't confuse it with props.

It will be state of this particular component so it accepts the previous state and this medaled will

modify the counter by 1 so it will use the previous state and substract one from the previous state

until it gets to zero.

Now our message might contain this counters so we can say you have confirmed your account and you'll

be redirected to the home page in and here you can render that counter.

So each second you will see that this timer goes down and down.

And also we need the previous state in component updates so that's the second parameter that gets passed

to them.

Campania did update the state component not the props.

And if the counter is different the previous counter value is different than the current one and that

it is less or equal zero.

Only then we will read direct to the home page.

So we also have this user register complete when component amounts and it will be a good idea to reset

this process when the counter goes to zero because this means everything was correct and everything

was fine.

So the counter went to zero user registered confirmed the token.

So that's the good good place to reset.

Also the state that we keep inside Rydex.

So let's go over it again.

So the confirmation success is different and it is true we initially set up the interval which will

call this function that sets the state to one less than the previous call each second.

So we initialize the state to 10 seconds.

Inside constructors so the subsequent calls to the said State will will reduce it by 1.

Now let's see how all of this looks in practice so also when the counter will go to zero.

We will do this finally direct and we control the components state.

Don't mistake with props inside the component did update because it's also passed the previous date

is also passed as the second parameter.

So make sure you have this e-mail that you are not registered with yet or delete the account with the

e-mail.

So now I got my confirmation talken I can copy it go back to the form paste the token and now you can

see that we have the counter and also inside the console.

The value is reducing.

And finally we are redirected to the home page.

So everything worked perfectly.