211.1. BlogPost form for creating posts and user permissions(Section 21: React Redux Application)
Now it's finally the time to let some of the users write new blog posts.

So we'll create a forum that allows users to create blog posts and also we'll check if they have divided

permissions for block post creation.

So let's start first with the blog posts forum so there is nothing new here.

As usual we import re-act their field component and relax form function.

We create our blog post forum extend re-act component wrap component first we've redox form gave the

form a name and finally use connect to connect it to our dispatching methods from redox.

So we have no because we aren't mapping and state

and.

For now let's simply render the create a new blog post so we have to show this blog post under a certain

route.

So obviously we have to define a new route for our forum so maybe let it be Blog Post forum and the

component would be blocked post Farm

case so we have our route defined to make sure that you imported the component inside the JS

And now the user is out indicated we should display a link for adding new blog posts so mind that this

will be currently shown to every user that is authenticated even to the ones that are commentators or

others that do not have rights to write a new blog post.

This is for example editors.

So now at least we see that it does work.

It renders that component so everything's fine here.

Now if I were to log in as someone who shouldn't have access to this it also would be displayed.

And the new user has the role commentary.

Commentator you can see that on the right on the network tab.

This user shouldn't have access to that blog post for him because if he would send this forum it would

be bug errors anyway so it might be a good idea to check if users have certain rights before letting

them send the data.