211.2. BlogPost form for creating posts and user permissions(Section 21: React Redux Application)
So we are getting only the user data from the substate of our global redox state.

So you do not have to always get the full subtree of your state you can for example get only one property

and will the helper methods for our insight our API utos and this would be called can write post.

So first we check the user data is not new.

And then we call some on the user data roles.

SOM is a method that you can provided a callback and on when called on the array.

It will check each of the elements if it is passing.

Check.

So if at least one element is passing a check it will return true.

So in our case we are iterating over all of the elements of this array and checking each one of them.

If it is one of the block both roles that we allow to write blog posts so it will check.

If at least one role that this user has because it might have many roles is either a roll writer role

admin or role super admin.

So we have the user data mapped to props from the state from the state of.

Exactly.

And we can check immediately if this user can write blog posts so even if he will land on this route

we can use our direct component.

So by returning returning a read direct component we will direct this user to the log in page.

If he tries to write a new blog post by for example entering the Eurail directly or later we'll also

fix it so the link is not shown for the users that do not have rights but now we can see that being

authenticated as a user who doesn't have rights it's still showing.

Hello €3 my name because well I registered always using my name but this was a different user and when

we check with an admin user you can see that it renders a form.