213.1. ImageUpload component (styling the file input)(Section 21: React Redux Application)
In this lecture will try to build an image uploading feature for our blog posts.

Let's start by logging in as an admin user and on the blog post for me you can see that we have a working

farm.

We have the validations and we are also checking each the user out Decatur it but we have a feature

of uploading files in the API but not in the re-act up yet.

So let's create an image upload J.S. component that will handle the image uploading.

So as usual we will import that they react to create clouds that will extend the re-act component and

this would be called the image upload.

And as always in our case this will extend the re-act component base class.

So for now we will only implement the Rander metalled So we will try to render the input of type file

to be able to select a file for uploading.

And since this component will be added to the blog post form we do not need any additional styling.

So the only required surrounding Deif class would have the form group class for properly formatting

the forms.

So we are adding the input of type file.

So this is for the file uploads and it should have the class name for control file which is the bootstrap

class for file input and also that text primary and found a way to bolt to modify the font.

That will be displayed.

You will see why we are doing this later because our final effect wouldn't include the traditional form

file input.

We would like it to be read read Stangl to which you can click or drag and drop the files.

So after exporting our image upload class we can now import it inside the blog post form and see if

we are able to render it.

So it looks well at least.

Well it doesn't really look well but at least it is being displayed.