213.2. ImageUpload component (styling the file input)(Section 21: React Redux Application)
So to actually modify it to the way we want it to look like we will create an image upload CSX to add

some styling to this element so it will look at the way we would like it to.

So the first step would be to add our custom class.

Let's call it nice image upload.

Doesn't really matter who can come up with a better name.

And this would be for the form control file that would be inside our div with that nice input upload

class.

So I notice I made an error.

I actually put this nice in boot upload class to the input element it should be in the surrounding Deif.

So we wanted to be 100 percent we've And hey if visibility would be hidden Kursaal we want it to be

pointer and the position we want it to be related to the normal position.

It would be rendered.

And we said the minimum head of this element to be sure it is big and visible.

And the trick we are going to do is we will before superclass for this element and this would be positioned

absolute which basically means this will be positioned relatively to its parent element.

And this is this time visible would also have the same width and we would like this rich Stangl to have

a border which will be two pixels and dasht color light gray.

And we want the border radius also we like that text to be aligned on the center

and let's maybe specify that the font color would be gray and font size could be the default.

Or you can make it bigger.

And now we specify the minimum hate us.

Also 6 and to render that text vertically on the center of vertical the center of this element will

add line hate us to and the padding top also asked you to a m.

So now let's move this class to the surrounding Deif element because otherwise this wouldn't actually

work.

And now we'll also do the trick to put this text that we want to that before subclass of this input

element and let it be Kleek me or drag and drop file.

And now inside this class we can use the content property and the other to specify which attribute of

this element should be used as content.

So now it doesn't work.

It haven't changed.

That's because we first have to import this CIRCE's file.

So let's do that now import the image upload see SS

fix that typo.

So this should be import.

And now after refreshing we should see our new shiny element file for file inputs and if you choose

a file it actually would be set to this element.

It's not uploading it will do that later but at least it works and you can drag and drop files on it

to.