214 Uploading images - upload request(Section 21: React Redux Application)
In this lecture we will continue working on our image upload component.

So we only render it nicely in the last lecture.

Now let's see how we can actually handle the file upload.

So we will have to specify on OWN change which would mean that the file has changed.

Someone selected a different file or dropped a different file on it and we use that this onchange being

this

to be able to access this in the context of this class which is the image upload.

And let's LOEG that target which would be to add element we can't access from that event and each target

files should contain the information about the file we have just added to this input element.

You can see some of the properties of this image like that.

Type the name of the file or even the last modified date.

So this is at least something.

Now we would directly beyond this component to dispatching our actions so let's import the conect function

from re-act redux.

And now let's define our constant map dispatched to prop's So we'll be able to dispatch an action.

This will be called image upload.

It doesn't exist here.

We'll be adding in this very lecture.

So first we of course have to export this component so let's use export default and then connect.

And we are not mapping any state so the first parameter can be no.

And the second should be our map dispatched to prop's object and wrap it with a call to with that image

upload.

Now also let's remove the export from that image upload because we only want this exported as a wrapped

component.

Now we need to create our auction creator.

So let's go to the actions G G-S and declare this function image upload.

And this will accept one parameter which would be called file and this would contain the file to upload.

So we'll also use the phunk middle were.

So we are returning function with one parameter dispatch from this very function.

And now we will call our mettled in our requests object.

So we currently only have post get and we should also add new upload.

So we are using it right now but we are yet to implement this method in the requests.

So we will be calling an end point as large images as it is for our image uploading.

We have one parameter file so when this is finished successfully we'll dispatch another action which

would be uploaded.

We of course have to create it.

And when something goes wrong we will dispatch image upload that image upload error.

But of course this has to be a function call so mind that I made a mistake.

It's not just the function name.

It has to be a function call to imagine upload that error.

Now we are defining Medhat upload inside our agent js file.

So we'll be see we'll see how we can handle uploads in super agent so to upload a file will send both

request to our to the point that is specified in the Eurail property.

But first will prepended with the API root bass.

Were all of our API.

And to upload a file we would call a patch.

So we are attaching a file which is passed of as an argument to this function to the field file.

So notice that this is how we defined our form for uploading images.

It has only one field which is file

now.

Also our token plug into it and then apply the response by the function to just get the actual response

body from the super agent.

We do not need anything else from it.

All right.

So this is how we will upload the files inside the super agent.

Now we have to implement those two action creators that image uploaded and image upload error to dispatch

them and later be able to handle those actions in our reducers.

So the first would be that you might upload it.

So let's return an object with that type image upload that it will contain the data of this image maybe

we can call it.

Well maybe you can call it just an image and pass the data to it so that our image or the data returned

from uploading that image will be available inside our producer.

And we will also image upload requests to somehow signal to our state that the process of uploading

and image has started.

So this would be useful to for example not let the users later submit their blog post form if the image

upload is in progress and we are also adding this image upload error.

And now we have to add this Konstanz to our Konstanz js file.

So let's start with that image.

Upload it.

And the next one would be that image upload request.

So you can just copy the first declaration whatever works for you whatever is faster we just need to

add this Konstanz.

So we have that image upload request and the last one is the image upload error or OK so they are all

defined.

And now finally we'll have to import those Konstanz inside actions.

Jay Yes but first let let's.

These bodged the image upload request action inside our image upload so we can in our reduce or we'll

be later able to figure out that that you match uploading is in progress.

So now we have imported this Konstanz by the way we can remove the no longer used block post list.

And reformat this list.

All right so our auction creator is done.

Now we can import it inside that image upload Geass file

and add the actual logic to to the on change event handler.

So whenever someone modifies the file that in the input of the type file or drugs and it drops this

file we can upload it.

So let's set Constand that will contain the data target files and the first element of that file because

we will only support uploading a single image at a time.

Now let's remove the import for this image upload and import it again because now it is our default

import we've redux.

So we have to import it differently.

So let's try uploading our first image and check on the network top if we actually have any requests

going.

So yes we do.

So there is that request to emerge as the first one is the options to figure out what methods are available

on the server.

Only the second one is the actual upload which returns the data of the uploaded image.

So if you see the admin panel you can see that those images uploaded.

Now let's maybe upload bunda.

So that would be it for this lecture.

Will continue in the next one.